+++ packages/client/hmi-client/src/components/workflow/util.ts
@@ -5,29 +5,6 @@ import type { WorkflowNode } from '@/types/workflow';
 import type { CalibrateMap } from '@/services/calibrate-workflow';
 import { useProjects } from '@/composables/project';
 
-/**
- * FIXME: This overlaps somewhat with services/dataset#createCsvAssetFromRunResults,
- *
- * This is a simpler version
- * - without dealing with a list/set of runResults, which is now depreated.
- * - no column stats, which are not used
- * */
-export const convertToCsvAsset = (data: Record<string, any>[], keys: string[]) => {
-	const csvData: CsvAsset = {
-		headers: keys,
-		csv: [],
-		rowCount: data.length
-	};
-	data.forEach((datum) => {
-		const row: any[] = [];
-		keys.forEach((k) => {
-			row.push(datum[k] || '');
-		});
-		csvData.csv.push(row);
-	});
-	return csvData;
-};
-
 // export const previewChartSize= (element: HTMLElement) => {
 // }
